pyproject.toml
[project]
name = "ryobi-gdo-2-mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "certifi>=2025.10.5",
    "ha-mqtt-discoverable>=0.22.1",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.12.0",
]

[project.scripts]
ryobi-gdo-2-mqtt = "ryobi_gdo_2_mqtt.__main__:main"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    # keep-sorted start
    "E",
    "F",
    "FURB",
    "G",
    "I",
    "ICN",
    "ISC",
    "LOG",
    "PIE",
    "PT",
    "Q",
    "RSE",
    "RUF",
    "T10",
    "UP",
    # keep-sorted end
]
ignore = [
    # keep-sorted start
    "E111",
    "E114",
    "E117",
    "E501",
    "E712",
    "ISC001",
    "PIE790",
    "Q001",
    "Q002",
    "Q003",
    "RUF005",
    "RUF006",
    "RUF012",
    "UP006",
    "UP007",
    "UP035",
    "UP045",
    "W191",
    # keep-sorted end
]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.basedpyright]
# keep-sorted start
exclude = ["**/node_modules", "**/__pycache__"]
include = ["src"]
pythonPlatform = "All"
pythonVersion = "3.13"
reportArgumentType = false
reportAttributeAccessIssue = false
reportIgnoreCommentWithoutRule = true
reportRedeclaration = false
reportUnnecessaryTypeIgnoreComment = "warning"
typeCheckingMode = "basic"
# keep-sorted end

executionEnvironments = [
    { root = "src/tests", reportPrivateUsage = false, extraPaths = [
        "src/tests/e2e",
    ] },
    { root = "src" },
]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
norecursedirs = [".git", "fixtures"]
log_format = "%(asctime)s.%(msecs)03d %(levelname)-8s %(threadName)s %(name)s:%(filename)s:%(lineno)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "auto"
addopts = [
    "--cov=ryobi_gdo_2_mqtt",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["src/ryobi_gdo_2_mqtt"]
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
